daemon xdp_queue: int8_t -> int16_t
I see little use in saving the bits here.  Our best benchmarking
machine has 128 threads and even 256 doesn't seem too far-fetched.
The card in there is documented to support even 1.5k queues (!)

+++ daemon/main.c
@@ -663,7 +663,7 @@ int main(int argc, char **argv)
 	}
 	/* Start listening on AF_XDP sockets. */
 	for (size_t i = 0; i < the_args->addrs_xdp.len; ++i) {
-		int8_t xdp_queue;
+		int16_t xdp_queue;
 		char *addr = the_args->addrs_xdp.at[i];
 		char *colon = strchr(addr, ':');
 		if (colon) {
